Split jisyo implementation
