+++ denops/skkeleton/jisyo.ts
@@ -1,25 +1,16 @@
 import { config } from "./config.ts";
-import { getKanaTable } from "./kana.ts";
-import { encoding } from "./deps/encoding_japanese.ts";
-import { wrap } from "./deps/iterator_helpers.ts";
 import { JpNum } from "./deps/japanese_numeral.ts";
 import { RomanNum } from "./deps/roman.ts";
 import { zip } from "./deps/std/collections.ts";
-import { TextLineStream } from "./deps/std/streams.ts";
-import { assert, is } from "./deps/unknownutil.ts";
-import { Encode } from "./types.ts";
-import { jisyoschema, jsonschema, msgpack, yaml } from "./deps/jisyo.ts";
-
-import type {
-  CompletionData,
-  Encoding,
-  RankData,
-  SkkServerOptions,
-} from "./types.ts";
+import type { CompletionData, RankData } from "./types.ts";
 import { readFileWithEncoding } from "./util.ts";
+import { SkkDictionary } from "./jisyo/skk_dictionary.ts";
+import { UserDictionary, UserDictionaryPath } from "./jisyo/user_dictionary.ts";
+import { SkkServer } from "./jisyo/skk_server.ts";
+import { GoogleJapaneseInput } from "./jisyo/google_japanese_input.ts";
 
-const okuriAriMarker = ";; okuri-ari entries.";
-const okuriNasiMarker = ";; okuri-nasi entries.";
+export const okuriAriMarker = ";; okuri-ari entries.";
+export const okuriNasiMarker = ";; okuri-nasi entries.";
 
 function toKifu(n: number): string {
   const x = Math.floor(n / 10);
@@ -122,14 +113,6 @@ export interface Dictionary {
   getCompletionResult(prefix: string, feed: string): Promise<CompletionData>;
 }
 
-function encode(str: string, encode: Encoding): Uint8Array {
-  const utf8Encoder = new TextEncoder();
-  const utf8Bytes = utf8Encoder.encode(str);
-  const eucBytesArray = encoding.convert(utf8Bytes, Encode[encode], "UTF8");
-  const eucBytes = Uint8Array.from(eucBytesArray);
-  return eucBytes;
-}
-
 export class NumberConvertWrapper implements Dictionary {
   #inner: Dictionary;
 
@@ -173,490 +156,7 @@ export function wrapDictionary(dict: Dictionary): Dictionary {
   );
 }
 
-interface Jisyo {
-  okuri_ari: Record<string, string[]>;
-  okuri_nasi: Record<string, string[]>;
-}
-
-export class SKKDictionary implements Dictionary {
-  #okuriAri: Map<string, string[]>;
-  #okuriNasi: Map<string, string[]>;
-
-  #cachedCandidates: Map<string, CompletionData>;
-
-  constructor(
-    okuriAri?: Map<string, string[]>,
-    okuriNasi?: Map<string, string[]>,
-  ) {
-    this.#okuriAri = okuriAri ?? new Map();
-    this.#okuriNasi = okuriNasi ?? new Map();
-    this.#cachedCandidates = new Map();
-  }
-
-  getHenkanResult(type: HenkanType, word: string): Promise<string[]> {
-    const target = type === "okuriari" ? this.#okuriAri : this.#okuriNasi;
-    return Promise.resolve(target.get(word) ?? []);
-  }
-
-  getCompletionResult(prefix: string, feed: string): Promise<CompletionData> {
-    const candidates: CompletionData = [];
-    if (feed != "") {
-      const table = getKanaTable();
-      for (const [key, kanas] of table) {
-        if (key.startsWith(feed) && kanas.length > 1) {
-          const feedPrefix = prefix + (kanas as string[])[0];
-          for (const entry of this.getCachedCandidates(prefix[0])) {
-            if (entry[0].startsWith(feedPrefix)) {
-              candidates.push(entry);
-            }
-          }
-        }
-      }
-    } else {
-      for (const entry of this.getCachedCandidates(prefix[0])) {
-        if (entry[0].startsWith(prefix)) {
-          candidates.push(entry);
-        }
-      }
-    }
-
-    candidates.sort((a, b) => a[0].localeCompare(b[0]));
-    return Promise.resolve(candidates);
-  }
-
-  private getCachedCandidates(prefix: string): CompletionData {
-    if (this.#cachedCandidates.has(prefix)) {
-      const candidates = this.#cachedCandidates.get(prefix);
-      return candidates ?? [];
-    }
-
-    const candidates: CompletionData = [];
-    for (const entry of this.#okuriNasi) {
-      if (entry[0].startsWith(prefix)) {
-        candidates.push(entry);
-      }
-    }
-
-    this.#cachedCandidates.set(prefix, candidates);
-    return candidates;
-  }
-
-  loadJson(data: string) {
-    const jisyo = JSON.parse(data) as Jisyo;
-    const validator = new jsonschema.Validator();
-    const result = validator.validate(jisyo, jisyoschema);
-    if (!result.valid) {
-      for (const error of result.errors) {
-        throw Error(error.message);
-      }
-    }
-    this.#okuriAri = new Map(Object.entries(jisyo.okuri_ari));
-    this.#okuriNasi = new Map(Object.entries(jisyo.okuri_nasi));
-  }
-
-  loadYaml(data: string) {
-    const jisyo = yaml.parse(data) as Jisyo;
-    const validator = new jsonschema.Validator();
-    const result = validator.validate(jisyo, jisyoschema);
-    if (!result.valid) {
-      for (const error of result.errors) {
-        throw Error(error.message);
-      }
-    }
-    this.#okuriAri = new Map(Object.entries(jisyo.okuri_ari));
-    this.#okuriNasi = new Map(Object.entries(jisyo.okuri_nasi));
-  }
-
-  loadMsgpack(data: Uint8Array) {
-    const jisyo = msgpack.decode(data) as Jisyo;
-    const validator = new jsonschema.Validator();
-    const result = validator.validate(jisyo, jisyoschema);
-    if (!result.valid) {
-      for (const error of result.errors) {
-        throw Error(error.message);
-      }
-    }
-    this.#okuriAri = new Map(Object.entries(jisyo.okuri_ari));
-    this.#okuriNasi = new Map(Object.entries(jisyo.okuri_nasi));
-  }
-
-  load(data: string) {
-    let mode = -1;
-    this.#okuriAri = new Map();
-    this.#okuriNasi = new Map();
-    const a: Map<string, string[]>[] = [this.#okuriAri, this.#okuriNasi];
-    const lines = data.split("\n");
-    for (const line of lines) {
-      if (line === okuriAriMarker) {
-        mode = 0;
-        continue;
-      }
-      if (line === okuriNasiMarker) {
-        mode = 1;
-        continue;
-      }
-      if (mode == -1) continue;
-      const pos = line.indexOf(" ");
-      if (pos !== -1) {
-        a[mode].set(line.substring(0, pos), line.slice(pos + 2, -1).split("/"));
-      }
-    }
-  }
-}
-
-type UserDictionaryPath = {
-  path?: string;
-  rankPath?: string;
-};
-
-export class UserDictionary implements Dictionary {
-  #okuriAri: Map<string, string[]>;
-  #okuriNasi: Map<string, string[]>;
-  #rank: Map<string, number>;
-
-  #path = "";
-  #rankPath = "";
-  #loadTime = -1;
-
-  #cachedPrefix = "";
-  #cachedFeed = "";
-  #cachedCandidates: CompletionData = [];
-
-  constructor(
-    okuriAri?: Map<string, string[]>,
-    okuriNasi?: Map<string, string[]>,
-    rank?: Map<string, number>,
-  ) {
-    this.#okuriAri = okuriAri ?? new Map();
-    this.#okuriNasi = okuriNasi ?? new Map();
-    this.#rank = rank ?? new Map();
-  }
-
-  getHenkanResult(type: HenkanType, word: string): Promise<string[]> {
-    const target = type === "okuriari" ? this.#okuriAri : this.#okuriNasi;
-    return Promise.resolve(target.get(word) ?? []);
-  }
-
-  private cacheCandidates(prefix: string, feed: string) {
-    if (this.#cachedPrefix === prefix && this.#cachedFeed == feed) {
-      return;
-    }
-    const candidates: CompletionData = [];
-    if (feed != "") {
-      const table = getKanaTable(config.kanaTable);
-      for (const [key, kanas] of table) {
-        if (key.startsWith(feed) && kanas.length > 1) {
-          const feedPrefix = prefix + (kanas as string[])[0];
-          for (const entry of this.#okuriNasi) {
-            if (entry[0].startsWith(feedPrefix)) {
-              candidates.push(entry);
-            }
-          }
-        }
-      }
-    } else {
-      for (const entry of this.#okuriNasi) {
-        if (entry[0].startsWith(prefix)) {
-          candidates.push(entry);
-        }
-      }
-    }
-    this.#cachedPrefix = prefix;
-    this.#cachedFeed = feed;
-    this.#cachedCandidates = candidates;
-  }
-
-  getCompletionResult(prefix: string, feed: string): Promise<CompletionData> {
-    this.cacheCandidates(prefix, feed);
-    return Promise.resolve(this.#cachedCandidates);
-  }
-
-  getRanks(prefix: string): RankData {
-    const set = new Set();
-    const adder = set.add.bind(set);
-    this.cacheCandidates(prefix, "");
-    for (const [, cs] of this.#cachedCandidates) {
-      cs.forEach(adder);
-    }
-    return wrap(this.#rank.entries())
-      .filter((e) => set.has(e[0]))
-      .toArray();
-  }
-
-  registerHenkanResult(type: HenkanType, word: string, candidate: string) {
-    if (candidate === "") {
-      return;
-    }
-    const target = type === "okuriari" ? this.#okuriAri : this.#okuriNasi;
-    const oldCandidate = target.get(word) ?? [];
-    target.set(
-      word,
-      Array.from(new Set([candidate, ...oldCandidate])),
-    );
-    this.#rank.set(candidate, Date.now());
-    this.#cachedPrefix = "";
-  }
-
-  purgeCandidate(type: HenkanType, word: string, candidate: string) {
-    const target = type === "okuriari" ? this.#okuriAri : this.#okuriNasi;
-    const newCandidate = (target.get(word) ?? []).filter((c) => c != candidate);
-    if (newCandidate.length > 0) {
-      target.set(word, newCandidate);
-    } else {
-      target.delete(word);
-    }
-  }
-
-  private async readFile(path: string, rankPath: string) {
-    let mode = -1;
-    this.#okuriAri = new Map();
-    this.#okuriNasi = new Map();
-    const a: Map<string, string[]>[] = [this.#okuriAri, this.#okuriNasi];
-    const lines = (await Deno.readTextFile(path)).split("\n");
-    for (const line of lines) {
-      if (line === okuriAriMarker) {
-        mode = 0;
-        continue;
-      }
-      if (line === okuriNasiMarker) {
-        mode = 1;
-        continue;
-      }
-      if (mode == -1) continue;
-      const pos = line.indexOf(" ");
-      if (pos !== -1) {
-        a[mode].set(line.substring(0, pos), line.slice(pos + 2, -1).split("/"));
-      }
-    }
-
-    // rank
-    if (!rankPath) {
-      return;
-    }
-    const rankData = JSON.parse(await Deno.readTextFile(rankPath));
-    assert(rankData, is.ArrayOf(is.String));
-    this.#rank = new Map(rankData.map((c, i) => [c, i]));
-  }
-
-  async load({ path, rankPath }: UserDictionaryPath = {}) {
-    path = this.#path = path ?? this.#path;
-    rankPath = this.#rankPath = rankPath ?? this.#rankPath;
-    if (path) {
-      try {
-        const stat = await Deno.stat(path);
-        const time = stat.mtime?.getTime() ?? -1;
-        if (time === this.#loadTime) {
-          return;
-        }
-        this.#loadTime = time;
-        await this.readFile(path, rankPath);
-      } catch {
-        // do nothing
-      }
-      this.#cachedPrefix = "";
-    }
-  }
-
-  private async writeFile(path: string, rankPath: string) {
-    // dictionary
-    // Note: in SKK dictionary reverses candidates sort order if okuriari
-    const okuriAri = Array.from(this.#okuriAri).sort((a, b) =>
-      b[0].localeCompare(a[0])
-    ).map((e) => `${e[0]} /${e[1].join("/")}/`);
-    const okuriNasi = Array.from(this.#okuriNasi).sort((a, b) =>
-      a[0].localeCompare(b[0])
-    ).map((e) => `${e[0]} /${e[1].join("/")}/`);
-    const data = [
-      [okuriAriMarker],
-      okuriAri,
-      [okuriNasiMarker],
-      okuriNasi,
-      [""],
-    ].flat().join("\n");
-    try {
-      await Deno.writeTextFile(path, data);
-    } catch (e) {
-      console.log(
-        `warning(skkeleton): can't write userJisyo to ${path}`,
-      );
-      throw e;
-    }
-    // rank
-    if (!rankPath) {
-      return;
-    }
-    const rankData = JSON.stringify(
-      Array.from(this.#rank.entries()).sort((a, b) => a[1] - b[1]).map((e) =>
-        e[0]
-      ),
-    );
-    try {
-      await Deno.writeTextFile(rankPath, rankData);
-    } catch (e) {
-      console.log(
-        `warning(skkeleton): can't write candidate rank data to ${rankPath}`,
-      );
-      throw e;
-    }
-  }
-
-  async save() {
-    if (!this.#path) {
-      return;
-    }
-    try {
-      await this.writeFile(this.#path, this.#rankPath);
-    } catch (e) {
-      if (config.debug) {
-        console.log(e);
-      }
-      return;
-    }
-    const stat = await Deno.stat(this.#path).catch(() => void 0);
-    this.#loadTime = stat?.mtime?.getTime() ?? -1;
-  }
-}
-
 export type HenkanType = "okuriari" | "okurinasi";
-
-async function* iterLine(
-  r: ReadableStream<Uint8Array>,
-  encoding: string,
-): AsyncIterable<string> {
-  const lines = r
-    .pipeThrough(new TextDecoderStream(encoding), {
-      preventAbort: true,
-      preventCancel: true,
-      preventClose: true,
-    })
-    .pipeThrough(new TextLineStream());
-
-  for await (const line of lines) {
-    if ((line as string).length) {
-      yield line as string;
-    }
-  }
-}
-
-export class SkkServer implements Dictionary {
-  #conn: Deno.Conn | undefined;
-  responseEncoding: Encoding;
-  requestEncoding: Encoding;
-  connectOptions: Deno.ConnectOptions;
-  constructor(opts: SkkServerOptions) {
-    this.requestEncoding = opts.requestEnc;
-    this.responseEncoding = opts.responseEnc;
-    this.connectOptions = {
-      hostname: opts.hostname,
-      port: opts.port,
-    };
-  }
-  async connect() {
-    this.#conn = await Deno.connect(this.connectOptions);
-  }
-  async getHenkanResult(_type: HenkanType, word: string): Promise<string[]> {
-    if (!this.#conn) return [];
-
-    await this.#conn.write(encode(`1${word} `, this.requestEncoding));
-    const result: string[] = [];
-    for await (
-      const str of iterLine(this.#conn.readable, this.responseEncoding)
-    ) {
-      result.push(...(str.at(0) === "4") ? [] : str.split("/").slice(1, -1));
-
-      if (str.endsWith("\n")) {
-        break;
-      }
-    }
-    return result;
-  }
-  async getCompletionResult(
-    prefix: string,
-    feed: string,
-  ): Promise<CompletionData> {
-    if (!this.#conn) return [];
-
-    let midashis: string[] = [];
-    if (feed != "") {
-      const table = getKanaTable();
-      for (const [key, kanas] of table) {
-        if (key.startsWith(feed) && kanas.length > 1) {
-          const feedPrefix = prefix + (kanas as string[])[0];
-          midashis = midashis.concat(await this.getMidashis(feedPrefix));
-        }
-      }
-    } else {
-      midashis = await this.getMidashis(prefix);
-    }
-
-    const candidates: CompletionData = [];
-    for (const midashi of midashis) {
-      candidates.push([
-        midashi,
-        await this.getHenkanResult("okurinasi", midashi),
-      ]);
-    }
-
-    return candidates;
-  }
-  private async getMidashis(prefix: string): Promise<string[]> {
-    // Get midashis from prefix
-    if (!this.#conn) return [];
-
-    await this.#conn.write(encode(`4${prefix} `, this.requestEncoding));
-    const midashis: string[] = [];
-    for await (
-      const str of iterLine(this.#conn.readable, this.responseEncoding)
-    ) {
-      midashis.push(...(str.at(0) === "4") ? [] : str.split("/").slice(1, -1));
-
-      if (str.endsWith("\n")) {
-        break;
-      }
-    }
-
-    return midashis;
-  }
-  close() {
-    this.#conn?.write(encode("0", this.requestEncoding));
-    this.#conn?.close();
-  }
-}
-
-export class GoogleJapaneseInput implements Dictionary {
-  async connect() {}
-  async getHenkanResult(_type: HenkanType, word: string): Promise<string[]> {
-    // It should not work for "okuriari".
-    return _type === "okuriari" ? [] : await this.getMidashis(word);
-  }
-  getCompletionResult(_prefix: string, _feed: string): Promise<CompletionData> {
-    // Note: It does not support completions
-    return Promise.resolve([]);
-  }
-  private async getMidashis(prefix: string): Promise<string[]> {
-    // Get midashis from prefix
-    const params = new URLSearchParams({
-      langpair: "ja-Hira|ja",
-      text: `${prefix},`,
-    });
-    try {
-      const resp = await fetch(
-        `http://www.google.com/transliterate?${params.toString()}`,
-        {
-          method: "GET",
-        },
-      );
-      const respJson = await resp.json();
-      return respJson[0][1];
-    } catch (e) {
-      if (config.debug) {
-        console.log(e);
-      }
-    }
-    return [];
-  }
-  close() {}
-}
 
 function gatherCandidates(
   collector: Map<string, Set<string>>,
@@ -671,7 +171,6 @@ function gatherCandidates(
 
 export class Library {
   #dictionaries: Dictionary[];
-
   #userDictionary: UserDictionary;
 
   constructor(
@@ -765,7 +264,7 @@ export async function load(
 ): Promise<Library> {
   const globalDictionaries = await Promise.all(
     globalDictionaryConfig.map(async ([path, encodingName]) => {
-      const dict = new SKKDictionary();
+      const dict = new SkkDictionary();
       try {
         if (path.endsWith(".yaml") || path.endsWith(".yml")) {
           const file = await Deno.readTextFile(path);
@@ -790,6 +289,7 @@ export async function load(
       return dict;
     }),
   );
+
   const userDictionary = new UserDictionary();
   try {
     await userDictionary.load(userDictionaryPath);
@@ -800,6 +300,7 @@ export async function load(
     }
     // do nothing
   }
+
   try {
     skkServer?.connect();
   } catch (e) {
@@ -808,6 +309,7 @@ export async function load(
       console.log(e);
     }
   }
+
   const dictionaries = globalDictionaries.map((d) => wrapDictionary(d));
   if (skkServer) {
     dictionaries.push(skkServer);
@@ -815,5 +317,6 @@ export async function load(
   if (googleJapaneseInput) {
     dictionaries.push(googleJapaneseInput);
   }
+
   return new Library(dictionaries, userDictionary);
 }
